+++ src/commons/VPagenation.tsx
@@ -1,4 +1,4 @@
-import PageButton from 'pages/profileList/PageButton';
+import PageButton from 'pages/ProfileList/PageButton';
 
 export function getPaginationItems(
   currentPage: number,
