+++ src/pages/ProfileList/urgentList/UrgentListPage.tsx
@@ -1,4 +1,4 @@
-import UrgentList from 'pages/profileList/urgentList/UrgentList';
+import UrgentList from 'pages/ProfileList/urgentList/UrgentList';
 import GNB from 'layouts/GNB';
 
 const UrgentListPage = () => {
