+++ src/pages/RegisterPage.tsx
@@ -1,8 +1,8 @@
 import DayModalInput from 'pages/register/DayModalInput';
 import MRegisterForm from 'pages/register/MRegisterForm';
 import GNB from 'layouts/GNB';
-import StatusSelectGroup from './StatusSelectGroup';
-import RegisterHeader from './RegisterHeader';
+import StatusSelectGroup from './register/StatusSelectGroup';
+import RegisterHeader from './register/RegisterHeader';
 
 const RegisterPage = () => {
   return (
